make CarFetcher more generic
