+++ gateway/backend_car_fetcher.go
@@ -7,14 +7,19 @@ import (
 	"io"
 	"math/rand"
 	"net/http"
+	"net/url"
+	"strconv"
+	"strings"
 	"time"
+
+	"github.com/ipfs/boxo/path"
 )
 
-type DataCallback func(resource string, reader io.Reader) error
+type DataCallback func(p path.ImmutablePath, reader io.Reader) error
 
 // CarFetcher powers a [CarBackend].
 type CarFetcher interface {
-	Fetch(ctx context.Context, path string, cb DataCallback) error
+	Fetch(ctx context.Context, path path.ImmutablePath, params CarParams, cb DataCallback) error
 }
 
 type remoteCarFetcher struct {
@@ -44,8 +49,10 @@ func NewRemoteCarFetcher(gatewayURL []string, httpClient *http.Client) (CarFetch
 	}, nil
 }
 
-func (ps *remoteCarFetcher) Fetch(ctx context.Context, path string, cb DataCallback) error {
-	urlStr := fmt.Sprintf("%s%s", ps.getRandomGatewayURL(), path)
+func (ps *remoteCarFetcher) Fetch(ctx context.Context, path path.ImmutablePath, params CarParams, cb DataCallback) error {
+	url := contentPathToCarUrl(path, params)
+
+	urlStr := fmt.Sprintf("%s%s", ps.getRandomGatewayURL(), url.String())
 	req, err := http.NewRequestWithContext(ctx, http.MethodGet, urlStr, nil)
 	if err != nil {
 		return err
@@ -79,6 +86,36 @@ func (ps *remoteCarFetcher) getRandomGatewayURL() string {
 	return ps.gatewayURL[ps.rand.Intn(len(ps.gatewayURL))]
 }
 
+// contentPathToCarUrl returns an URL that allows retrieval of specified resource
+// from a trustless gateway that implements IPIP-402
+func contentPathToCarUrl(path path.ImmutablePath, params CarParams) *url.URL {
+	return &url.URL{
+		Path:     path.String(),
+		RawQuery: carParamsToString(params),
+	}
+}
+
+// carParamsToString converts CarParams to URL parameters compatible with IPIP-402
+func carParamsToString(params CarParams) string {
+	paramsBuilder := strings.Builder{}
+	paramsBuilder.WriteString("format=car") // always send explicit format in URL, this  makes debugging easier, even when Accept header was set
+	if params.Scope != "" {
+		paramsBuilder.WriteString("&dag-scope=")
+		paramsBuilder.WriteString(string(params.Scope))
+	}
+	if params.Range != nil {
+		paramsBuilder.WriteString("&entity-bytes=")
+		paramsBuilder.WriteString(strconv.FormatInt(params.Range.From, 10))
+		paramsBuilder.WriteString(":")
+		if params.Range.To != nil {
+			paramsBuilder.WriteString(strconv.FormatInt(*params.Range.To, 10))
+		} else {
+			paramsBuilder.WriteString("*")
+		}
+	}
+	return paramsBuilder.String()
+}
+
 type retryCarFetcher struct {
 	inner   CarFetcher
 	retries int
@@ -99,12 +136,12 @@ func NewRetryCarFetcher(inner CarFetcher, allowedRetries int) (CarFetcher, error
 	}, nil
 }
 
-func (r *retryCarFetcher) Fetch(ctx context.Context, path string, cb DataCallback) error {
-	return r.fetch(ctx, path, cb, r.retries)
+func (r *retryCarFetcher) Fetch(ctx context.Context, path path.ImmutablePath, params CarParams, cb DataCallback) error {
+	return r.fetch(ctx, path, params, cb, r.retries)
 }
 
-func (r *retryCarFetcher) fetch(ctx context.Context, path string, cb DataCallback, retriesLeft int) error {
-	err := r.inner.Fetch(ctx, path, cb)
+func (r *retryCarFetcher) fetch(ctx context.Context, path path.ImmutablePath, params CarParams, cb DataCallback, retriesLeft int) error {
+	err := r.inner.Fetch(ctx, path, params, cb)
 	if err == nil {
 		return nil
 	}
@@ -124,8 +161,8 @@ func (r *retryCarFetcher) fetch(ctx context.Context, path string, cb DataCallbac
 		// Resets the number of retries for partials, mimicking Caboose logic.
 		retriesLeft = r.retries
 
-		return r.fetch(ctx, t.StillNeed[0], cb, retriesLeft)
+		return r.fetch(ctx, t.StillNeed[0].Path, t.StillNeed[0].Params, cb, retriesLeft)
 	default:
-		return r.fetch(ctx, path, cb, retriesLeft)
+		return r.fetch(ctx, path, params, cb, retriesLeft)
 	}
 }
